build: remove unnecessary build time operations
Avoid copying the modules everywhere.  This was cargo-culted from
swift-models without the understanding that it was done for demo
purposes rather than for build reasons.  The unnecessary duplication
slows the build, increases the output directory and is just generally
unnecessary.

Use the proper variable to identify the C compiler rather than some
arbitrary path which does not exist.  Avoid incorrect variables being
defined.