build: import CCUDA and SwiftRTCuda from the SwiftRT build
These libraries are required for building SwiftRT with the CUDA backend
and using that to build models.  This unfortunately requires some
additional setup locally as we want to build SwiftRT as an external
component rather than consume an existing previous build via CMake's
`-D [PROJECT]_DIR` approach which would allow us to evade some of the
local setup.